refactor: move termData-independent helpers into termHelpers
extract `parseLocalDate`, `parseQuarter`, `termSchema`, and
`canTermEnrollmentChange` from `term.ts` into `termHelpers.ts`. these
four don't read the `termData` const at runtime, so they don't need to
live in the module that loads `$generated/termData.json` (and the
`$components` chain via the `CourseEvent` import).

`term.ts` re-imports `parseQuarter` + `termSchema` for its own use.
`PlannerStore` and `get-search-data` still need `termData`-coupled
stuff from `term.ts`, so they keep a `$lib/term` import for that but
pull the helpers from `$lib/termHelpers` directly.
