+++ apps/antalmanac/scripts/get-search-data.ts
@@ -2,7 +2,8 @@ import 'dotenv/config';
 import { access, mkdir, writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
 
-import { canTermEnrollmentChange, termData } from '$lib/term';
+import { termData } from '$lib/term';
+import { canTermEnrollmentChange } from '$lib/termHelpers';
 import type { AATerm, CourseSearchResult, DepartmentSearchResult } from '@packages/antalmanac-types';
 import { createClient } from '@packages/anteater-api/client';
 import type { Course, WebsocAPIResponse, WebsocCourse, WebsocDepartment } from '@packages/anteater-api/types';
